5_semestre/sistemas_distribuidos/anomaly-detection-kafka-streamlit/requirements.txt
numpy==1.22.4
pandas==1.5.3
streamlit==1.22.0
kafka-python==2.0.2
plotly==5.13.0
matplotlib==3.7.1
seaborn==0.12.2